+++ src/main/java/com/unimag/tiendauniversitaria/api/dto/ProductDtos.java
@@ -2,23 +2,25 @@ package com.unimag.tiendauniversitaria.api.dto;
 
 import jakarta.validation.constraints.*;
 
+import java.math.BigDecimal;
+
 public class ProductDtos {
     public record ProductCreateRequest(
             @NotBlank String name,
-            @NotNull Double price,
+            @NotNull BigDecimal price,
             @NotNull Long categoryId
     ) {}
 
     public record ProductUpdateRequest(
             @NotBlank String name,
-            @NotNull Double price,
+            @NotNull BigDecimal price,
             @NotNull Long categoryId
     ) {}
 
     public record ProductResponse(
             Long id,
             String name,
-            Double price,
+            BigDecimal price,
             String categoryName
     ) {}
 }
